requirements.txt
# Core Distributed Processing
pyspark==3.5.0
delta-spark==3.1.0

# Data Handling & Utilities
pandas==2.2.0
pyyaml==6.0.1

# Telemetry Generator
faker==19.13.0

# Monitoring & Dashboards
plotly==5.18.0

# Notebooks
jupyterlab==4.1.0

# Testing
pytest==8.0.0
pytest-mock==3.12.0

# Optional: Linting & Formatting
black==24.1.1
ruff==0.2.0
